requirements.txt
mido==1.3.3
pandas==2.3.3
matplotlib==3.10.6
PyYAML==6.0.3
numpy==2.3.3
Unidecode==1.4.0
torch
tqdm
pretty_midi